+++ mayan/settings/production.py
@@ -8,11 +8,11 @@ DEBUG = False
 # https://docs.djangoproject.com/en/dev/ref/settings/#allowed-hosts
 ALLOWED_HOSTS = ['*']
 
-TEMPLATE_LOADERS = (
-    ('django.template.loaders.cached.Loader', (
+TEMPLATES['OPTIONS']['loaders'] = (
+    'django.template.loaders.cached.Loader', (
         'django.template.loaders.filesystem.Loader',
         'django.template.loaders.app_directories.Loader',
-    )),
+    )
 )
 
 CELERY_ALWAYS_EAGER = False